+++ litebird_sim/__init__.py
@@ -1,10 +1,10 @@
 # -*- encoding: utf-8 -*-
-from pathlib import Path
 
 import numba
 
 from litebird_sim.mapmaking import (
     make_binned_map,
+    make_brahmap_gls_map,
     check_valid_splits,
     BinnerResult,
     make_destriped_map,
@@ -38,6 +38,7 @@ from .coordinates import (
     GAL_TO_ECL_EULER,
     CoordinateSystem,
     coord_sys_to_healpix_string,
+    rotate_coordinates_e2g,
 )
 from .detectors import (
     DetectorInfo,
@@ -73,6 +74,7 @@ from .hwp_sys.hwp_sys import (
     HwpSys,
 )
 from .imo import (
+    PTEP_IMO_LOCATION,
     Imo,
     FormatSpecification,
     Entity,
@@ -84,7 +86,7 @@ from .io import (
     read_list_of_observations,
 )
 from .madam import save_simulation_for_madam
-from .mbs.mbs import Mbs, MbsParameters, MbsSavedMapInfo
+from .mbs.mbs import FG_MODELS, Mbs, MbsParameters, MbsSavedMapInfo
 from .mpi import MPI_COMM_WORLD, MPI_ENABLED, MPI_CONFIGURATION, MPI_COMM_GRID
 from .mueller_convolver import MuellerConvolver
 from .noise import (
@@ -181,8 +183,6 @@ from .version import __author__, __version__
 # work queue
 numba.config.THREADING_LAYER_CONFIG = ["tbb", "omp", "workqueue"]
 
-PTEP_IMO_LOCATION = Path(__file__).parent.parent / "default_imo"
-
 
 __all__ = [
     "__author__",
@@ -235,6 +235,7 @@ __all__ = [
     # madam.py
     "save_simulation_for_madam",
     # mbs.py
+    "FG_MODELS",
     "Mbs",
     "MbsParameters",
     "MbsSavedMapInfo",
@@ -299,6 +300,7 @@ __all__ = [
     "apply_hwp_to_obs",
     # mapmaking
     "make_binned_map",
+    "make_brahmap_gls_map",
     "check_valid_splits",
     "BinnerResult",
     "make_destriped_map",
@@ -332,6 +334,7 @@ __all__ = [
     "GAL_TO_ECL_EULER",
     "CoordinateSystem",
     "coord_sys_to_healpix_string",
+    "rotate_coordinates_e2g",
     # spacecraft.py
     "compute_l2_pos_and_vel",
     "compute_lissajous_pos_and_vel",